finished adding a friend to the list
