+++ friends/src/App.js
@@ -3,7 +3,6 @@ import Friends from './component/FriendsPage';
 import Login from './component/Login';
 import {Route, Link} from 'react-router-dom';
 import {PrivateRoute} from './component/PrivateRoute';
-import AddFriends from './component/AddFriends'
 
 import './App.css';
 
@@ -13,10 +12,8 @@ function App() {
       <h1>Insta-Friends</h1>
       <Link to='/'>Login</Link>
       <Link to='/friends'>Friends List</Link>
-      <Link to='/friends/add'>Add Friends</Link>
       <Route exact path='/'component={Login}></Route>
-      <PrivateRoute exact path='/friends'component={Friends}/>
-      <PrivateRoute exact path='/friends/add'component={AddFriends}/>
+      <PrivateRoute path='/friends'component={Friends}/>
     </div>
   );
 }
